+++ sdk/communication/azure-communication-chat/CHANGELOG.md
@@ -18,8 +18,9 @@
      - Added `PushNotificationCallback` as the type for a customized push notification handler.
      - Added instructions in README.md in sample app for steps to setup FCM push notifications.
      - Added code in sample app for FCM push notification end to end scenario.
-- Added `startRealtimeNotifications(Context context)` to `ChatAsyncClient`.
-- Added `startRealtimeNotifications(Context context)` to `ChatClient`.
+- Added `startRealtimeNotifications(Context context, Consumer<Throwable> errorHandler)` to `ChatAsyncClient`.
+- Added `startRealtimeNotifications(Context context, Consumer<Throwable> errorHandler)` to `ChatClient`.
+- Updated `com.azure.android.core` from `1.0.0-beta.6` to `1.0.0-beta.7`.
 
 ## 1.1.0-beta.2 (2021-08-10)
 - Version update.